+++ workflow/src/test/kotlin/com/r3/corda/sdk/token/workflow/MockNetworkTest.kt
@@ -4,6 +4,8 @@ import co.paralleluniverse.fibers.Suspendable
 import com.r3.corda.sdk.token.contracts.states.EvolvableTokenType
 import com.r3.corda.sdk.token.contracts.types.TokenType
 import com.r3.corda.sdk.token.contracts.utilities.withNotary
+import com.r3.corda.sdk.token.workflow.flows.ConfidentialIssueFlow
+import com.r3.corda.sdk.token.workflow.flows.ConfidentialMoveFlow
 import com.r3.corda.sdk.token.workflow.flows.CreateEvolvableToken
 import com.r3.corda.sdk.token.workflow.flows.IssueToken
 import com.r3.corda.sdk.token.workflow.flows.MoveToken
@@ -162,56 +164,5 @@ abstract class MockNetworkTest(val names: List<CordaX500Name>) {
                 amount = amount,
                 anonymous = anonymous
         ))
-    }
-
-    object ConfidentialIssueFlow {
-        @InitiatingFlow
-        class Initiator<T : TokenType>(
-                val token: T,
-                val holder: Party,
-                val notary: Party,
-                val amount: Amount<T>? = null
-        ) : FlowLogic<SignedTransaction>() {
-            @Suspendable
-            override fun call(): SignedTransaction {
-                val holderSession = initiateFlow(holder)
-                val confidentialHolder = subFlow(RequestConfidentialIdentity.Initiator(holderSession)).party.anonymise()
-                return subFlow(IssueToken.Initiator(token, confidentialHolder, notary, amount, holderSession))
-            }
-        }
-
-        @InitiatedBy(Initiator::class)
-        class Responder(val otherSession: FlowSession) : FlowLogic<Unit>() {
-            @Suspendable
-            override fun call() {
-                subFlow(RequestConfidentialIdentity.Responder(otherSession))
-                subFlow(IssueToken.Responder(otherSession))
-            }
-        }
-    }
-
-    object ConfidentialMoveFlow {
-        @InitiatingFlow
-        class Initiator<T : TokenType>(
-                val ownedToken: T,
-                val holder: Party,
-                val amount: Amount<T>? = null
-        ) : FlowLogic<SignedTransaction>() {
-            @Suspendable
-            override fun call(): SignedTransaction {
-                val holderSession = initiateFlow(holder)
-                val confidentialHolder = subFlow(RequestConfidentialIdentity.Initiator(holderSession)).party.anonymise()
-                return subFlow(MoveToken.Initiator(ownedToken, confidentialHolder, amount, holderSession))
-            }
-        }
-
-        @InitiatedBy(Initiator::class)
-        class Responder(val otherSession: FlowSession) : FlowLogic<Unit>() {
-            @Suspendable
-            override fun call() {
-                subFlow(RequestConfidentialIdentity.Responder(otherSession))
-                subFlow(MoveToken.Responder(otherSession))
-            }
-        }
     }
 }
